+++ ParallaxGenLib/src/patchers/PatcherUpgradeParallaxToCM.cpp
@@ -6,9 +6,12 @@
 #include "patchers/PatcherShader.hpp"
 
 #include <filesystem>
+#include <mutex>
 #include <utility>
 
 using namespace std;
+
+mutex PatcherUpgradeParallaxToCM::UpgradeCMMutex;
 
 auto PatcherUpgradeParallaxToCM::getFactory() -> PatcherShaderTransform::PatcherShaderTransformFactory {
   return [](filesystem::path NIFPath, nifly::NifFile *NIF) -> PatcherShaderTransformObject {
@@ -26,6 +29,8 @@ PatcherUpgradeParallaxToCM::PatcherUpgradeParallaxToCM(std::filesystem::path NIF
 
 auto PatcherUpgradeParallaxToCM::transform(const PatcherShader::PatcherMatch &FromMatch)
     -> PatcherShader::PatcherMatch {
+  lock_guard<mutex> Lock(UpgradeCMMutex);
+
   const auto HeightMap = FromMatch.MatchedPath;
 
   // Get texture base (remove _p.dds)